Bug 4976: Fix NPE in SimpleMapCache

+++ mappingservice/implementation/src/main/java/org/opendaylight/lispflowmapping/implementation/mapcache/SimpleMapCache.java
@@ -244,7 +244,7 @@ public class SimpleMapCache implements IMapCache {
                 }
             } else {
                 Date timestamp = (Date) daoEntry.getValue().get(SubKeys.REGDATE);
-                if (MappingMergeUtil.timestampIsExpired(timestamp)) {
+                if (timestamp != null && MappingMergeUtil.timestampIsExpired(timestamp)) {
                     dao.removeSpecific(daoEntry.getKey(), SubKeys.REGDATE);
                     dao.removeSpecific(daoEntry.getKey(), SubKeys.RECORD);
                     return null;
